requirements.txt
Django==4.2.7
djongo==1.3.6
pymongo==3.12.3
Pillow==10.1.0
dnspython==2.4.2
